# openapi.yaml
openapi: 3.0.0
info:
  title: Actifai Engineering Takehome API
  version: "1.0.0"
  description: API for retrieving user and group sales data, including revenue calculations.

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Confirms the API is running.
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: Hello World!

  /users/all:
    get:
      summary: Retrieve all users
      description: Returns a list of all users in the system.
      responses:
        '200':
          description: List of user objects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    name:
                      type: string
                      example: Alice

  /users/{id}:
    get:
      summary: Retrieve a user by ID
      description: Returns details for a specific user.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A user object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: Alice
        '404':
          description: User not found

  /users/{id}/averageRevenue:
    get:
      summary: Get a user's average revenue
      description: Returns the average revenue for a user over a specified date range (or all-time if no date is specified).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date in YYYY-MM-DD format.
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date in YYYY-MM-DD format.
      responses:
        '200':
          description: The user's average revenue
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    example: 11
                  averageRevenue:
                    type: number
                    format: float
                    example: 1500.25

  /users/{id}/averageDailyRevenue:
    get:
      summary: Get a user's average daily revenue
      description: Returns a time series of the user's daily average revenue within a specified date range (or all-time).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date in YYYY-MM-DD format.
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date in YYYY-MM-DD format.
      responses:
        '200':
          description: List of daily average revenue values
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      format: date
                      example: "2021-08-06"
                    averageSale:
                      type: number
                      format: float
                      example: 3341

  /users/{id}/averageMonthlyRevenue:
    get:
      summary: Get a user's average monthly revenue
      description: Returns a time series of the user's monthly average revenue within a specified date range (or all-time).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date in YYYY-MM-DD format.
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date in YYYY-MM-DD format.
      responses:
        '200':
          description: List of monthly average revenue values
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      format: date
                      example: "2021-08"
                    averageSale:
                      type: number
                      format: float
                      example: 4000.75

  /groups/all:
    get:
      summary: Retrieve all groups
      description: Returns a list of all groups.
      responses:
        '200':
          description: List of group objects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    name:
                      type: string
                      example: "Sales Team"

  /groups/{id}:
    get:
      summary: Retrieve a group by ID
      description: Returns details for a specific group.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A group object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: "Sales Team"
        '404':
          description: Group not found

  /groups/{id}/averageRevenue:
    get:
      summary: Get a group's average revenue
      description: Returns the average revenue for a group over a specified date range (or all-time if not provided).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date in YYYY-MM-DD format.
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date in YYYY-MM-DD format.
      responses:
        '200':
          description: The group's average revenue
          content:
            application/json:
              schema:
                type: object
                properties:
                  group_id:
                    type: integer
                    example: 2
                  averageRevenue:
                    type: number
                    format: float
                    example: 8000.50

  /groups/{id}/averageDailyRevenue:
    get:
      summary: Get a user's average daily revenue
      description: Returns a time series of the user's daily average revenue within a specified date range (or all-time).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date in YYYY-MM-DD format.
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date in YYYY-MM-DD format.
      responses:
        '200':
          description: List of daily average revenue values
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      format: date
                      example: "2021-08-06"
                    averageSale:
                      type: number
                      format: float
                      example: 3341

  /groups/{id}/averageMonthlyRevenue:
    get:
      summary: Get a user's average monthly revenue
      description: Returns a time series of the user's monthly average revenue within a specified date range (or all-time).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date in YYYY-MM-DD format.
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date in YYYY-MM-DD format.
      responses:
        '200':
          description: List of monthly average revenue values
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      format: date
                      example: "2021-08"
                    averageSale:
                      type: number
                      format: float
                      example: 4000.75

  /sales/all:
    get:
      summary: Retrieve all sales records
      description: Returns a list of all sales transactions.
      responses:
        '200':
          description: List of sales records
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 19
                    user_id:
                      type: integer
                      example: 20
                    amount:
                      type: number
                      format: float
                      example: 3341
                    date:
                      type: string
                      format: date
                      example: "2021-08-06"